+++ examples/simple-native/src/main.rs
@@ -6,7 +6,7 @@ use fuel_indexer::{
 };
 use fuels::core::{abi_decoder::ABIDecoder, ParamType, Tokenizable};
 
-use fuel_indexer_derive::graphql_schema;
+use fuel_indexer_macros::graphql_schema;
 use fuels_abigen_macro::abigen;
 
 use std::path::PathBuf;
