+++ fuel-indexer-macros/src/lib.rs
