+++ fuel-indexer-macros/test_data/badschema.graphql
