+++ fuel-indexer-macros/test_data/fail_args.rs
@@ -2,13 +2,16 @@
 extern crate alloc;
 use alloc::vec::Vec;
 use core::convert::TryFrom;
-use fuel_indexer_derive::handler;
-use fuel_indexer_plugins::types::*;
+use fuel_indexer_macros::handler;
+use fuel_indexer_plugin::types::*;
 use fuels_core::{ParamType, Token};
 
-struct Logger;
-impl Logger {
-    pub fn info(_: &str) {}
+#[handler(SomeArg)]
+fn function_one(event: SomeEvent) {
+    let SomeEvent { id, account } = event;
+
+    let t1 = Thing1 { id, account };
+    t1.save();
 }
 
 struct SomeEvent {
