+++ fuel-indexer-macros/test_data/fail_args.stderr
