+++ fuel-indexer-macros/test_data/fail_noschema.rs
@@ -2,8 +2,8 @@
 extern crate alloc;
 use alloc::vec::Vec;
 use core::convert::TryFrom;
-use fuel_indexer_derive::{graphql_schema, handler};
-use fuel_indexer_plugins::types::*;
+use fuel_indexer_macros::{graphql_schema, handler};
+use fuel_indexer_plugin::types::*;
 use fuels_core::{ParamType, Token};
 
 struct Logger;
