+++ fuel-indexer-macros/test_data/fail_noschema.stderr
