+++ fuel-indexer-macros/test_data/fail_self.stderr
