+++ fuel-indexer-macros/test_data/schema.graphql
