+++ fuel-indexer-plugin/src/lib.rs
